Moves FormFieldNumberTable
