+++ src/components/Form/Col/Col.tsx
@@ -13,15 +13,15 @@ import FormFieldImageUpload from "@root/components/Field/FormFieldImageUpload";
 import FormFieldImageVideoLinkDocumentBrowsing from "@root/components/Field/FormFieldImageVideoLinkDocumentBrowsing";
 import FormFieldMapCoordinates from "@root/components/Field/FormFieldMapCoordinates";
 import FormFieldMatrix from "@root/components/Field/FormFieldMatrix";
-import FormFieldText from "@root/forms/FormFieldText";
+import FormFieldNumberTable from "@root/components/Field/FormFieldNumberTable";
 import FormFieldPhoneSelectionDropdown from "@root/forms/FormFieldPhoneSelectionDropdown";
 import FormFieldRadio from "@root/forms/FormFieldRadio";
 import FormFieldRaw from "@root/forms/FormFieldRaw";
 import FormFieldToggleSwitch from "@root/forms/FormFieldToggleSwitch";
 import FormFieldTable from "@root/forms/FormFieldTable";
+import FormFieldText from "@root/forms/FormFieldText";
 import FormFieldTextEditor from "@root/forms/FormFieldTextEditor";
 import FormFieldUpload from "@root/forms/FormFieldUpload";
-import FormFieldNumberTable from "@root/forms/FormFieldNumberTable";
 
 import evaluateShow from "@root/utils/show/evaluateShow";
 import RegisteredField from "../../Field/RegisteredField";
